Fix CategoryModel mapping in CategoriesRepositoryImplTest
The `id` property of `GenreCacheModel` is an `Int`, so direct mapping to `CategoryModel`'s `id` (also `Int`) is now used, removing the unnecessary `toInt()` conversion.

+++ repository/search/src/test/kotlin/com/repository/search/repository/CategoriesRepositoryImplTest.kt
@@ -43,7 +43,7 @@ class CategoriesRepositoryImplTest {
         val result = repository.getAllCategories()
 
         // Then
-        assertEquals(localGenres.map { CategoryModel(it.id.toInt(), it.name) }, result)
+        assertEquals(localGenres.map { CategoryModel(it.id, it.name) }, result)
         coVerify(exactly = 0) { genresRemoteDataSource.getAllGenres() }
     }
